# docs/swagger.yaml
definitions:
  server.CreateWalletReq:
    properties:
      coinName:
        description: 币种名称
        type: string
      protocol:
        description: 协议
        type: string
    required:
    - coinName
    - protocol
    type: object
  server.CreateWalletRes:
    properties:
      address:
        description: 生成的钱包地址
        type: string
    type: object
  server.DelWalletReq:
    properties:
      address:
        description: 地址
        type: string
      coinName:
        description: 币种名称
        type: string
      protocol:
        description: 协议
        type: string
    required:
    - address
    - coinName
    - protocol
    type: object
  server.Response:
    properties:
      code:
        description: 错误code码
        type: integer
      data:
        description: 成功时返回的对象
      message:
        description: 错误信息
        type: string
    type: object
  server.TransactionReceiptReq:
    properties:
      coinName:
        description: 币种名称
        type: string
      hash:
        description: 交易哈希
        type: string
      protocol:
        description: 协议
        type: string
    required:
    - coinName
    - hash
    - protocol
    type: object
  server.TransactionReceiptRes:
    properties:
      status:
        description: 交易状态（0：未成功，1：已成功）
        type: integer
    type: object
  server.WithdrawReq:
    properties:
      address:
        description: 提现地址
        type: string
      coinName:
        description: 币种名称
        type: string
      orderId:
        description: 订单号
        type: string
      protocol:
        description: 协议
        type: string
      value:
        description: 金额
        type: integer
    required:
    - address
    - coinName
    - orderId
    - protocol
    - value
    type: object
  server.WithdrawRes:
    properties:
      hash:
        description: 生成的交易hash
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /api/createWallet:
    post:
      parameters:
      - description: 参数
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/server.CreateWalletReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.Response'
            - properties:
                data:
                  $ref: '#/definitions/server.CreateWalletRes'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建钱包地址
      tags:
      - 钱包
  /api/delWallet:
    post:
      parameters:
      - description: 参数
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/server.DelWalletReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除钱包地址
      tags:
      - 钱包
  /api/getTransactionReceipt:
    get:
      parameters:
      - description: 参数
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/server.TransactionReceiptReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.Response'
            - properties:
                data:
                  $ref: '#/definitions/server.TransactionReceiptRes'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取交易结果
      tags:
      - 钱包
  /api/withdraw:
    post:
      parameters:
      - description: 参数
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/server.WithdrawReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.Response'
            - properties:
                data:
                  $ref: '#/definitions/server.WithdrawRes'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 提现
      tags:
      - 钱包
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"